feat: move db message structs into model package
